backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
pydantic

# Testing dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
faker==33.1.0
pytest-mock==3.14.0
